@@ -1,11 +1,12 @@
 {
-  "name": "recompose-build",
+  "name": "recompose",
   "private": true,
   "author": "Andrew Clark <acdlite@me.com>",
   "repository": {
     "type": "git",
     "url": "git://github.com/acdlite/recompose.git"
   },
+  "version": "0.30.1-ts",
   "license": "MIT",
   "scripts": {
     "lint": "eslint scripts src",
@@ -13,10 +14,11 @@
     "test": "jest && flow check && cross-env SNAPSHOT=match npm run build:recompose",
     "test:watch": "cross-env BABEL_ENV=cjs jest --watch",
     "release": "node scripts/release.js",
-    "postinstall": "node scripts/installNestedPackageDeps.js",
+    "__postinstall": "node scripts/installNestedPackageDeps.js",
     "format": "prettier --semi false --trailing-comma es5 --single-quote --write 'scripts/*.js' 'src/packages/*/*.js' 'src/packages/*/!(node_modules)/**/*.js'",
     "precommit": "lint-staged",
-    "prepush": "yarn test"
+    "prepush": "yarn test",
+    "prepare": "cross-env NODE_ENV=cjs node_modules/.bin/babel ./src/packages/recompose --out-dir ./lib/packages/recompose --ignore='**/__tests__/**,**/node_modules/**'"
   },
   "jest": {
     "testMatch": [
@@ -26,6 +28,7 @@
       "text-summary",
       "lcov"
     ],
+    "testURL": "http://localhost/",
     "setupTestFrameworkScriptFile": "<rootDir>/scripts/jest.setup.js"
   },
   "lint-staged": {
@@ -43,6 +46,7 @@
     "@babel/preset-env": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
     "@babel/runtime": "^7.0.0",
+    "@wojtekmaj/enzyme-adapter-react-17": "0.8.0",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^9.0.0",
     "babel-jest": "^22.4.3",
@@ -69,19 +73,19 @@
     "most": "^1.0.2",
     "prettier": "^1.2.2",
     "prop-types": "^15.6.1",
-    "react": "^16.3.1",
-    "react-dom": "^16.3.1",
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0",
     "readline-sync": "^1.2.21",
     "rollup": "^0.65.0",
     "rollup-plugin-babel": "^4.0.1",
     "rollup-plugin-commonjs": "^9.1.6",
     "rollup-plugin-node-resolve": "^3.3.0",
     "rollup-plugin-replace": "^2.0.0",
-    "rollup-plugin-size-snapshot": "^0.6.1",
+    "rollup-plugin-size-snapshot": "^0.12.0",
     "rollup-plugin-uglify": "^4.0.0",
     "rx": "^4.1.0",
     "rxjs": "^5.0.0",
-    "shelljs": "^0.6.0",
+    "shelljs": "^0.8.5",
     "sinon": "^1.17.1",
     "webpack": "^2.4.1",
     "xstream": "^5.0.5"
@@ -90,7 +94,5 @@
     "node": "5.x",
     "npm": "3.x"
   },
-  "dependencies": {
-    "enzyme-adapter-react-16": "^1.1.1"
-  }
+  "dependencies": {}
 }
