@@ -4,7 +4,7 @@ import { lifecycle } from '../'
 
 test('lifecycle is a higher-order component version of React.Component', () => {
   const enhance = lifecycle({
-    componentWillMount() {
+    UNSAFE_componentWillMount() {
       this.setState({ 'data-bar': 'baz' })
     },
   })
